chore: refactor all report modules live
