--- lib/pescarte_web/live/pesquisa/relatorio/index.html.heex
@@ -1,11 +1,0 @@
-<.live_component
-  module={PescarteWeb.Pesquisa.Relatorio.FormComponent}
-  id={@relatorio.id_publico || :new}
-  report_type={@report_type}
-  title={@title}
-  action={@live_action}
-  relatorio={@relatorio}
-  today={@today}
-  pesquisador_id={@pesquisador_id}
-  patch={~p"/app/pesquisa/relatorios"}
-/>
