--- lib/pescarte_web/live/pesquisa/relatorio/list_report_live.html.heex
@@ -1,40 +1,0 @@
-<div class="search">
-  <.search_input name="busca-relatorio" id="busca-relatorio" />
-
-  <div class="links-item" style="display: flex;">
-    <.button class="icon-button" style="primary" click="preencher-relatorio" value="mensal">
-      <Lucideicons.book class="text-white-100" />
-      <.text size="sm" color="text-white-100">Relatório Mensal</.text>
-    </.button>
-  </div>
-
-  <div class="links-item">
-    <.button class="icon-button" style="primary" submit>
-      <Lucideicons.book class="text-white-100" />
-      <.text size="sm" color="text-white-100">Preencher Relatório</.text>
-    </.button>
-  </div>
-
-  <div :if={@current_user.tipo == :admin} class="links-item">
-    <.button class="icon-button" style="primary" submit>
-      <Lucideicons.files class="text-white-100" />
-      <.text size="sm" color="text-white-100">Compilar</.text>
-    </.button>
-  </div>
-</div>
-
-<.table rows={@relatorios} text-color="text-blue-100">
-  <:column :let={relatorios} label="  ">
-    <div class="flex-center"><.checkbox name="check" id="usuario" /></div>
-  </:column>
-  <:column :let={relatorios} label="Data"><%= relatorios.data %></:column>
-  <:column :let={relatorios} label="Tipo" class="linhas"><%= relatorios.tipo %></:column>
-  <:column :let={relatorios} label="Mês/Ano">
-    <%= relatorios.mes <> " / " <> relatorios.ano %>
-  </:column>
-  <:column :let={relatorios} label="Nome"><%= relatorios.name %></:column>
-  <:column :let={relatorios} label="Status"><%= relatorios.status %></:column>
-  <:column :let={relatorios} label="Baixar">
-    <div class="flex-center"><Lucideicons.download /></div>
-  </:column>
-</.table>
