+++ lib/pescarte_web/live/pesquisa/relatorio_live/show.html.heex
@@ -1,0 +1,8 @@
+<.live_component
+  module={PescarteWeb.Pesquisa.RelatorioLive.FormComponent}
+  id={@relatorio.id_publico}
+  title={@page_title}
+  action={@live_action}
+  relatorio={@relatorio}
+  patch={~p"/app/pesquisa/relatorios"}
+/>
